requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.1.4
django-allauth==65.3.1
django-cleanup==9.0.0
django-crispy-forms==2.3
django-filter==24.3
django-htmx==1.21.0
django-humanize==0.1.2
django-jazzmin==3.0.1
django-mathfilters==1.0.0
Faker==33.1.0
humanize==4.11.0
idna==3.10
lxml==5.3.0
pillow==11.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.2
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
